requirements.txt
async-timeout      4.0.3
beautifulsoup4     4.12.3
certifi            2024.2.2
charset-normalizer 3.3.2
dnspython          2.6.1
idna               3.7
lxml               5.2.1
mongoengine        0.28.2
pip                24.0
pymongo            4.7.0
python-dotenv      1.0.1
redis              5.0.4
redis-lru          0.1.2
requests           2.31.0
setuptools         57.4.0
soupsieve          2.5
urllib3            2.2.1